backend/requirements.txt
fastapi==0.115.11
jose==1.0.0
numpy==2.2.4
pandas==2.2.3
pydantic==2.10.6
python_bcrypt==0.3.2
Requests==2.32.3
scikit_learn==1.6.1
SQLAlchemy==2.0.39
